+++ kolkoKrzyzyk/GridLayout.hpp
@@ -6,7 +6,7 @@ class GridLayout
 public:
 	GridLayout(int startX, int startY, int endX, int endY);
 	void setRows(int horizontalRows, int verticalRows);
-	sf::Vector2i getPosition(int vertRow, int horRow);
+	sf::Vector2f getPosition(int vertRow, int horRow);
 private:
 	int startX;
 	int startY;
